neurad-studio/requirements.txt
absl-py==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array_record==0.8.1
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
av==15.0.0
av2==0.2.1
babel==2.17.0
backcall==0.2.0
beartype==0.17.2
beautifulsoup4==4.13.4
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
cmake==4.1.0
colorlog==6.9.0
comm==0.2.3
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
dask==2023.3.1
dataclass-wizard==0.35.1
dataclass_array==1.5.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dependency-groups==1.3.1
descartes==1.1.0
dill==0.4.0
distlib==0.4.0
dm-tree==0.1.9
dnspython==2.7.0
docstring_parser==0.17.0
einops==0.7.0
einsum==0.3.0
email_validator==2.2.0
embreex==2.17.7.post6
entrypoints==0.4
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.8
fastjsonschema==2.21.2
filelock==3.19.1
fire==0.7.1
flatbuffers==25.2.10
fonttools==4.59.1
fqdn==1.5.1
fsspec==2025.7.0
gast==0.4.0
gmplot==1.4.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.74.0
gsplat==1.5.3
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
immutabledict==2.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.37.0
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
jaxtyping==0.2.38
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==2.11.0
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.15.2
lit==18.1.8
llvmlite==0.44.0
locket==1.0.0
lxml==6.0.0
manifold3d==3.2.1
mapbox_earcut==1.0.3
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
mistune==3.1.3
mpmath==1.3.0
msgspec==0.19.0
multiprocess==0.70.18
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nerfacc==0.5.3
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nodeenv==1.9.1
notebook==7.4.5
notebook_shim==0.2.4
nox==2025.5.1
numba==0.61.2
numpy==1.26.4
numpy-quaternion==2024.0.10
nuscenes-devkit==1.1.11
oauthlib==3.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
OpenEXR==1.3.9
opt_einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandas==1.5.3
-e git+https://github.com/scaleapi/pandaset-devkit.git@59be180e2a3f3e37f6d66af9e67bf944ccbf6ec0#egg=pandaset&subdirectory=python
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathos==0.3.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==4.3.8
plotly==5.13.1
pox==0.3.6
ppft==1.7.7
prometheus_client==0.22.1
promise==2.3
prompt_toolkit==3.0.51
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==10.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.0
pycocotools==2.0.10
pycollada==0.9.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyproj==3.7.1
pyquaternion==0.9.9
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-msssim==1.0.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==27.0.2
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rich-toolkit==0.15.0
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.3
Shapely==1.8.5.post1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
splines==0.3.3
stack-data==0.6.3
starlette==0.37.2
svg.path==7.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.13.0
tensorflow-probability==0.19.0
termcolor==3.1.0
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn.git@3f6ef958edd4c5ab4d894b6e4813839cc809457c#subdirectory=bindings/torch
toml==0.10.2
tomli==2.2.1
toolz==0.12.0
#torch==2.0.1+cu118
#torch==2.0.1
torchmetrics==1.1.2
#torchvision==0.15.2+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.7.4
triton==2.0.0
typeguard==2.13.3
typer==0.16.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.14.1
tyro==0.7.3
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
vhacdx==0.0.8.post2
virtualenv==20.34.0
viser==1.0.6
visu3d==1.5.1
wadler_lindig==0.1.7
watchfiles==1.1.0
#waymo-open-dataset-tf-2-11-0==1.6.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
xxhash==3.5.0
yourdfpy==0.0.58
zipp==3.23.0
zod==0.8.0
